requirements.txt
gradio>=4.0.0
pillow
numpy
scikit-learn
matplotlib
scipy
opencv-python-headless
opencv-contrib-python-headless
